backend/requirements.txt
# RightCut backend dependencies
# All free — no paid APIs required

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9  # required for file uploads
websockets>=12.0

# ── LLM ───────────────────────────────────────────────────────────────────────
google-genai>=1.0.0      # Gemini 2.5 Flash — free tier via Google AI Studio

# ── Excel engine ──────────────────────────────────────────────────────────────
openpyxl>=3.1.5

# ── Document parsing ───────────────────────────────────────────────────────────
pypdf>=4.0.0             # PDF text extraction (fallback)
pdfplumber>=0.11.0       # PDF table extraction (preferred)
python-docx>=1.1.0       # DOCX parsing
pandas>=2.2.0            # CSV / XLSX reading
openpyxl>=3.1.5          # pandas Excel engine (already above)

# ── Config & validation ────────────────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0

# ── Dev / testing (optional) ──────────────────────────────────────────────────
# httpx>=0.27.0           # for async test client
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
